AI | Split ai-mcp toolsets into per-feature mcp submodules
