+++ modules/flexibleSearch/mcp/resources/META-INF/sap.commerce.toolset-flexibleSearch-mcp.xml
@@ -19,13 +19,7 @@
 <idea-plugin>
 
     <extensions defaultExtensionNs="com.intellij">
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.GroovyMcpToolset"/>
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.FlexibleSearchMcpToolset"/>
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.ImpExMcpToolset"/>
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.HacMcpToolset"/>
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.SolrMcpToolset"/>
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.LoggingMcpToolset"/>
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.ai.mcp.TypeSystemMcpToolset"/>
+        <mcpServer.mcpToolset implementation="sap.commerce.toolset.flexibleSearch.mcp.FlexibleSearchMcpToolset"/>
     </extensions>
 
 </idea-plugin>
