+++ modules/groovy/mcp/build.gradle.kts
@@ -35,9 +35,10 @@ sourceSets {
 
 dependencies {
     implementation(project(":shared-core"))
-    implementation(project(":exec-core"))
     implementation(project(":hac-exec"))
-    implementation(project(":solr-exec"))
+    implementation(project(":groovy-exec"))
+    implementation(project(":exec-core"))
+    implementation(project(":ai-mcp"))
 
     intellijPlatform {
         intellijIdea(properties("intellij.version")) {
@@ -49,4 +50,3 @@ dependencies {
         )
     }
 }
-
