+++ modules/hac/mcp/build.gradle.kts
@@ -37,7 +37,7 @@ dependencies {
     implementation(project(":shared-core"))
     implementation(project(":exec-core"))
     implementation(project(":hac-exec"))
-    implementation(project(":solr-exec"))
+    implementation(project(":ai-mcp"))
 
     intellijPlatform {
         intellijIdea(properties("intellij.version")) {
@@ -49,4 +49,3 @@ dependencies {
         )
     }
 }
-
