+++ resources/META-INF/plugin.xml
@@ -154,7 +154,13 @@
     <depends optional="true" config-file="sap.commerce.toolset-database-core.xml">com.intellij.database</depends>
     <depends optional="true" config-file="sap.commerce.toolset-cron-core.xml">com.intellij.cron</depends>
     <depends optional="true" config-file="sap.commerce.toolset-javascript-core.xml">JavaScript</depends>
-    <depends optional="true" config-file="sap.commerce.toolset-ai-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-hac-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-groovy-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-flexibleSearch-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-impex-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-solr-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-logging-mcp.xml">com.intellij.mcpServer</depends>
+    <depends optional="true" config-file="sap.commerce.toolset-typeSystem-mcp.xml">com.intellij.mcpServer</depends>
 
     <xi:include href="/META-INF/plugin-internal.xml" xpointer="xpointer(/idea-plugin/*)"/>
 </idea-plugin>
